:recycle: Refactor : 유저 정보 요청 작업 고도화
